backend/requirements.txt
# Core backend framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.1

# Supabase integration
supabase>=2.3.3

# Data validation
pydantic>=2.8.2

# Environment variable management
python-dotenv>=1.0.1

# Date and time handling
python-dateutil>=2.9.0.post0

# Optional (for scheduling notifications if you expand later)
APScheduler>=3.10.4

# CORS and async utilities
starlette>=0.38.2
httpx>=0.25.0

# Firebase Cloud Messaging
firebase-admin>=6.0.0

#langchain
langchain>=0.1.0
langchain-google-genai>=0.0.6